+++ src/tools.cpp
@@ -46,15 +46,15 @@ MatrixXd Tools::CalculateJacobian(const VectorXd &x_state) {
 
   MatrixXd Hj(3, 4);
   // recover state parameters
-  double px = x_state(0);
-  double py = x_state(1);
-  double vx = x_state(2);
-  double vy = x_state(3);
+  float px = x_state(0);
+  float py = x_state(1);
+  float vx = x_state(2);
+  float vy = x_state(3);
 
   // pre-compute a set of terms to avoid repeated calculation
-  double c1 = pow(px, 2) + pow(py, 2);
-  double c2 = sqrt(c1);
-  double c3 = (c1 * c2);
+  float c1 = pow(px, 2) + pow(py, 2);
+  float c2 = sqrt(c1);
+  float c3 = (c1 * c2);
 
   // check division by zero
   if (fabs(c1) < 0.0001) {
